// src/src2/api/models/LigneBackOrder.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LigneBackOrder = {
    idProduit?: string;
    referenceProduit?: string;
    nomProduit?: string;
    unite?: string;
    quantiteCommandee?: number;
    quantiteRecue?: number;
    quantiteEnAttente?: number;
    prixUnitaire?: number;
    montantTotal?: number;
    notes?: string;
};